requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
Django==3.1.2
django-cors-headers==3.5.0
django-premailer==0.2.0
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
lxml==4.6.1
Pillow==8.0.0
premailer==3.0.0
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
